app/requirements.txt
# ejecutar en la terminal pip install -r app\requirements.txt, para instalar todos lo necesarios para el proyecto

blinker==1.7.0
click==8.1.7
Flask==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
mysql-connector-python==8.3.0
Werkzeug==3.0.1
